[3d] refactor load3d nodes
