+++ src/extensions/core/load3d/Load3d.ts
@@ -1,71 +1,40 @@
 import { LGraphNode } from '@comfyorg/litegraph'
 import * as THREE from 'three'
-import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls'
-import { ViewHelper } from 'three/examples/jsm/helpers/ViewHelper'
-import { FBXLoader } from 'three/examples/jsm/loaders/FBXLoader'
-import { GLTF, GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader'
-import { MTLLoader } from 'three/examples/jsm/loaders/MTLLoader'
-import { OBJLoader } from 'three/examples/jsm/loaders/OBJLoader'
-import { STLLoader } from 'three/examples/jsm/loaders/STLLoader'
 
-import Load3dUtils from '@/extensions/core/load3d/Load3dUtils'
-import { useToastStore } from '@/stores/toastStore'
-
-interface Load3DOptions {
-  createPreview?: boolean
-  node?: LGraphNode
-}
+import { CameraManager } from './CameraManager'
+import { ControlsManager } from './ControlsManager'
+import { EventManager } from './EventManager'
+import { LightingManager } from './LightingManager'
+import { LoaderManager } from './LoaderManager'
+import { ModelManager } from './ModelManager'
+import { NodeStorage } from './NodeStorage'
+import { PreviewManager } from './PreviewManager'
+import { SceneManager } from './SceneManager'
+import { ViewHelperManager } from './ViewHelperManager'
+import {
+  CameraState,
+  CaptureResult,
+  Load3DOptions,
+  MaterialMode,
+  UpDirection
+} from './interfaces'
 
 class Load3d {
-  scene: THREE.Scene
-  perspectiveCamera: THREE.PerspectiveCamera
-  orthographicCamera: THREE.OrthographicCamera
-  activeCamera: THREE.Camera
   renderer: THREE.WebGLRenderer
-  controls: OrbitControls
-  gltfLoader: GLTFLoader
-  objLoader: OBJLoader
-  mtlLoader: MTLLoader
-  fbxLoader: FBXLoader
-  stlLoader: STLLoader
-  currentModel: THREE.Object3D | null = null
-  originalModel: THREE.Object3D | THREE.BufferGeometry | GLTF | null = null
-  animationFrameId: number | null = null
-  gridHelper: THREE.GridHelper
-  lights: THREE.Light[] = []
-  clock: THREE.Clock
-  normalMaterial: THREE.MeshNormalMaterial
-  standardMaterial: THREE.MeshStandardMaterial
-  wireframeMaterial: THREE.MeshBasicMaterial
-  depthMaterial: THREE.MeshDepthMaterial
-  originalMaterials: WeakMap<THREE.Mesh, THREE.Material | THREE.Material[]> =
-    new WeakMap()
+  protected clock: THREE.Clock
+  protected animationFrameId: number | null = null
+  protected node: LGraphNode
 
-  materialMode: 'original' | 'normal' | 'wireframe' | 'depth' = 'original'
-  currentUpDirection: 'original' | '-x' | '+x' | '-y' | '+y' | '-z' | '+z' =
-    'original'
-  originalRotation: THREE.Euler | null = null
-  viewHelper: ViewHelper = {} as ViewHelper
-  viewHelperContainer: HTMLDivElement = {} as HTMLDivElement
-  previewRenderer: THREE.WebGLRenderer | null = null
-  previewCamera: THREE.Camera | null = null
-  previewContainer: HTMLDivElement = {} as HTMLDivElement
-  targetWidth: number = 1024
-  targetHeight: number = 1024
-  showPreview: boolean = true
-  previewWidth: number = 120
-  node: LGraphNode = {} as LGraphNode
-  listeners: { [key: string]: Function[] } = {}
-
-  backgroundScene: THREE.Scene
-  backgroundCamera: THREE.OrthographicCamera
-  backgroundMesh: THREE.Mesh | null = null
-  backgroundTexture: THREE.Texture | null = null
-
-  previewBackgroundScene: THREE.Scene
-  previewBackgroundCamera: THREE.OrthographicCamera
-  previewBackgroundMesh: THREE.Mesh | null = null
-  previewBackgroundTexture: THREE.Texture | null = null
+  protected eventManager: EventManager
+  protected nodeStorage: NodeStorage
+  protected sceneManager: SceneManager
+  protected cameraManager: CameraManager
+  protected controlsManager: ControlsManager
+  protected lightingManager: LightingManager
+  protected viewHelperManager: ViewHelperManager
+  protected previewManager: PreviewManager
+  protected loaderManager: LoaderManager
+  protected modelManager: ModelManager
 
   constructor(
     container: Element | HTMLElement,
@@ -74,1115 +43,248 @@ class Load3d {
     }
   ) {
     this.node = options.node || ({} as LGraphNode)
-    this.scene = new THREE.Scene()
-
-    this.perspectiveCamera = new THREE.PerspectiveCamera(75, 1, 0.1, 1000)
-    this.perspectiveCamera.position.set(5, 5, 5)
-
-    const frustumSize = 10
-    this.orthographicCamera = new THREE.OrthographicCamera(
-      -frustumSize / 2,
-      frustumSize / 2,
-      frustumSize / 2,
-      -frustumSize / 2,
-      0.01,
-      10000
-    )
-    this.orthographicCamera.position.set(5, 5, 5)
-
-    this.activeCamera = this.perspectiveCamera
-
-    this.perspectiveCamera.lookAt(0, 0, 0)
-    this.orthographicCamera.lookAt(0, 0, 0)
+    this.clock = new THREE.Clock()
 
     this.renderer = new THREE.WebGLRenderer({ alpha: true, antialias: true })
     this.renderer.setSize(300, 300)
     this.renderer.setClearColor(0x282828)
     this.renderer.autoClear = false
-
     this.renderer.outputColorSpace = THREE.SRGBColorSpace
+    container.appendChild(this.renderer.domElement)
 
-    const rendererDomElement: HTMLCanvasElement = this.renderer.domElement
+    this.eventManager = new EventManager()
+    this.nodeStorage = new NodeStorage(this.node)
 
-    container.appendChild(rendererDomElement)
-
-    this.controls = new OrbitControls(
-      this.activeCamera,
-      this.renderer.domElement
+    this.sceneManager = new SceneManager(
+      this.renderer,
+      this.getActiveCamera.bind(this),
+      this.getControls.bind(this),
+      this.eventManager
     )
-    this.controls.enableDamping = true
 
-    this.controls.addEventListener('end', () => {
-      this.storeNodeProperty('Camera Info', this.getCameraState())
-    })
+    this.cameraManager = new CameraManager(
+      this.renderer,
+      this.eventManager,
+      this.nodeStorage
+    )
 
-    this.gltfLoader = new GLTFLoader()
-    this.objLoader = new OBJLoader()
-    this.mtlLoader = new MTLLoader()
-    this.fbxLoader = new FBXLoader()
-    this.stlLoader = new STLLoader()
-    this.clock = new THREE.Clock()
+    this.controlsManager = new ControlsManager(
+      this.renderer,
+      this.cameraManager.activeCamera,
+      this.eventManager,
+      this.nodeStorage
+    )
 
-    this.setupLights()
+    this.cameraManager.setControls(this.controlsManager.controls)
 
-    this.gridHelper = new THREE.GridHelper(10, 10)
-    this.gridHelper.position.set(0, 0, 0)
-    this.scene.add(this.gridHelper)
+    this.lightingManager = new LightingManager(
+      this.sceneManager.scene,
+      this.eventManager
+    )
 
-    this.normalMaterial = new THREE.MeshNormalMaterial({
-      flatShading: false,
-      side: THREE.DoubleSide,
-      normalScale: new THREE.Vector2(1, 1),
-      transparent: false,
-      opacity: 1.0
-    })
+    this.viewHelperManager = new ViewHelperManager(
+      this.renderer,
+      this.getActiveCamera.bind(this),
+      this.getControls.bind(this),
+      this.nodeStorage
+    )
 
-    this.wireframeMaterial = new THREE.MeshBasicMaterial({
-      color: 0xffffff,
-      wireframe: true,
-      transparent: false,
-      opacity: 1.0
-    })
+    this.previewManager = new PreviewManager(
+      this.sceneManager.scene,
+      this.getActiveCamera.bind(this),
+      this.getControls.bind(this),
+      () => this.renderer,
+      this.eventManager,
+      this.sceneManager.backgroundScene,
+      this.sceneManager.backgroundCamera
+    )
 
-    this.depthMaterial = new THREE.MeshDepthMaterial({
-      depthPacking: THREE.BasicDepthPacking,
-      side: THREE.DoubleSide
-    })
+    this.modelManager = new ModelManager(
+      this.sceneManager.scene,
+      this.renderer,
+      this.eventManager,
+      this.getActiveCamera.bind(this),
+      this.setupCamera.bind(this)
+    )
 
-    this.standardMaterial = this.createSTLMaterial()
+    this.loaderManager = new LoaderManager(this.modelManager, this.eventManager)
 
-    this.createViewHelper(container)
+    this.sceneManager.init()
+    this.cameraManager.init()
+    this.controlsManager.init()
+    this.lightingManager.init()
+    this.loaderManager.init()
+    this.loaderManager.init()
+
+    this.viewHelperManager.createViewHelper(container)
+    this.viewHelperManager.init()
 
     if (options && options.createPreview) {
-      this.createCapturePreview(container)
+      this.previewManager.createCapturePreview(container)
+      this.previewManager.init()
     }
 
-    this.backgroundScene = new THREE.Scene()
-    this.backgroundCamera = new THREE.OrthographicCamera(-1, 1, 1, -1, -1, 1)
-    this.previewBackgroundScene = this.backgroundScene.clone()
-    this.previewBackgroundCamera = this.backgroundCamera.clone()
-
-    const planeGeometry = new THREE.PlaneGeometry(2, 2)
-    const planeMaterial = new THREE.MeshBasicMaterial({
-      transparent: true,
-      depthWrite: false,
-      depthTest: false,
-      side: THREE.DoubleSide
-    })
-
-    this.backgroundMesh = new THREE.Mesh(planeGeometry, planeMaterial)
-    this.backgroundMesh.position.set(0, 0, 0)
-
-    this.previewBackgroundMesh = this.backgroundMesh.clone()
-
-    this.backgroundScene.add(this.backgroundMesh)
-    this.previewBackgroundScene.add(this.previewBackgroundMesh)
-
     this.handleResize()
-
     this.startAnimation()
   }
 
-  updateBackgroundSize(
-    backgroundTexture: THREE.Texture | null,
-    backgroundMesh: THREE.Mesh | null,
-    targetWidth: number,
-    targetHeight: number
-  ) {
-    if (!backgroundTexture || !backgroundMesh) return
-
-    const material = backgroundMesh.material as THREE.MeshBasicMaterial
-
-    if (!material.map) return
-
-    const imageAspect =
-      backgroundTexture.image.width / backgroundTexture.image.height
-    const targetAspect = targetWidth / targetHeight
-
-    if (imageAspect > targetAspect) {
-      backgroundMesh.scale.set(imageAspect / targetAspect, 1, 1)
-    } else {
-      backgroundMesh.scale.set(1, targetAspect / imageAspect, 1)
-    }
-
-    material.needsUpdate = true
+  private getActiveCamera(): THREE.Camera {
+    return this.cameraManager.activeCamera
   }
 
-  async setBackgroundImage(uploadPath: string) {
-    if (uploadPath === '') {
-      this.removeBackgroundImage()
-      return
-    }
-
-    let imageUrl = Load3dUtils.getResourceURL(
-      ...Load3dUtils.splitFilePath(uploadPath)
-    )
-
-    if (!imageUrl.startsWith('/api')) {
-      imageUrl = '/api' + imageUrl
-    }
-
-    try {
-      const textureLoader = new THREE.TextureLoader()
-      const texture = await new Promise<THREE.Texture>((resolve, reject) => {
-        textureLoader.load(imageUrl, resolve, undefined, reject)
-      })
-
-      if (this.backgroundTexture) {
-        this.backgroundTexture.dispose()
-      }
-
-      if (this.previewBackgroundTexture) {
-        this.previewBackgroundTexture.dispose()
-      }
-
-      texture.colorSpace = THREE.SRGBColorSpace
-
-      this.backgroundTexture = texture
-      this.previewBackgroundTexture = texture
-
-      const material = this.backgroundMesh?.material as THREE.MeshBasicMaterial
-      material.map = texture
-      material.needsUpdate = true
-
-      const material2 = this.previewBackgroundMesh
-        ?.material as THREE.MeshBasicMaterial
-      material2.map = texture
-      material2.needsUpdate = true
-
-      this.backgroundMesh?.position.set(0, 0, 0)
-      this.previewBackgroundMesh?.position.set(0, 0, 0)
-
-      this.updateBackgroundSize(
-        this.previewBackgroundTexture,
-        this.previewBackgroundMesh,
-        this.targetWidth,
-        this.targetHeight
-      )
-
-      this.emitEvent('backgroundImageChange', uploadPath)
-    } catch (error) {
-      console.error('Error loading background image:', error)
-    }
+  private getControls() {
+    return this.controlsManager.controls
   }
 
-  removeBackgroundImage() {
-    if (this.backgroundMesh) {
-      const material = this.backgroundMesh.material as THREE.MeshBasicMaterial
-      material.map = null
-      material.needsUpdate = true
-    }
-
-    if (this.previewBackgroundMesh) {
-      const material2 = this.previewBackgroundMesh
-        .material as THREE.MeshBasicMaterial
-      material2.map = null
-      material2.needsUpdate = true
-    }
-
-    if (this.backgroundTexture) {
-      this.backgroundTexture.dispose()
-      this.backgroundTexture = null
-    }
-
-    if (this.previewBackgroundTexture) {
-      this.previewBackgroundTexture.dispose()
-      this.previewBackgroundTexture = null
-    }
-
-    this.renderer.render(this.scene, this.activeCamera)
-    if (this.previewRenderer && this.previewCamera) {
-      this.previewRenderer.render(this.scene, this.previewCamera)
-    }
+  private setupCamera(size: THREE.Vector3): void {
+    this.cameraManager.setupForModel(size)
   }
 
-  addEventListener(event: string, callback: Function) {
-    if (!this.listeners[event]) {
-      this.listeners[event] = []
-    }
-    this.listeners[event].push(callback)
-  }
-
-  removeEventListener(event: string, callback: Function) {
-    if (this.listeners[event]) {
-      this.listeners[event] = this.listeners[event].filter(
-        (cb) => cb !== callback
-      )
-    }
-  }
-
-  emitEvent(event: string, data?: any) {
-    if (this.listeners[event]) {
-      this.listeners[event].forEach((callback) => callback(data))
-    }
-  }
-
-  storeNodeProperty(name: string, value: any) {
-    if (this.node) {
-      this.node.properties[name] = value
-    }
-  }
-
-  loadNodeProperty(name: string, defaultValue: any) {
-    if (
-      !this.node ||
-      !this.node.properties ||
-      !(name in this.node.properties)
-    ) {
-      return defaultValue
-    }
-    return this.node.properties[name]
-  }
-
-  createCapturePreview(container: Element | HTMLElement) {
-    this.previewRenderer = new THREE.WebGLRenderer({
-      alpha: true,
-      antialias: true,
-      preserveDrawingBuffer: true
-    })
-    this.previewRenderer.setSize(this.targetWidth, this.targetHeight)
-    this.previewRenderer.setClearColor(0x282828)
-    this.previewRenderer.autoClear = false
-
-    this.previewRenderer.outputColorSpace = THREE.SRGBColorSpace
-
-    this.previewContainer = document.createElement('div')
-    this.previewContainer.style.cssText = `
-      position: absolute;
-      right: 0;
-      bottom: 0;
-      background: rgba(0, 0, 0, 0.2);
-      display: block;
-      transition: border-color 0.1s ease;
-    `
-    this.previewContainer.appendChild(this.previewRenderer.domElement)
-
-    const MIN_PREVIEW_WIDTH = 120
-    const MAX_PREVIEW_WIDTH = 240
-
-    this.previewContainer.addEventListener('wheel', (event) => {
-      event.preventDefault()
-      event.stopPropagation()
-
-      const delta = event.deltaY
-      const oldWidth = this.previewWidth
-
-      if (delta > 0) {
-        this.previewWidth = Math.max(MIN_PREVIEW_WIDTH, this.previewWidth - 10)
-      } else {
-        this.previewWidth = Math.min(MAX_PREVIEW_WIDTH, this.previewWidth + 10)
-      }
-
-      if (
-        oldWidth !== this.previewWidth &&
-        (this.previewWidth === MIN_PREVIEW_WIDTH ||
-          this.previewWidth === MAX_PREVIEW_WIDTH)
-      ) {
-        this.flashPreviewBorder()
-      }
-
-      this.updatePreviewSize()
-      this.updatePreviewRender()
-    })
-
-    this.previewContainer.style.display = this.showPreview ? 'block' : 'none'
-
-    container.appendChild(this.previewContainer)
-  }
-
-  flashPreviewBorder() {
-    const originalBorder = this.previewContainer.style.border
-    const originalBoxShadow = this.previewContainer.style.boxShadow
-
-    this.previewContainer.style.border = '2px solid rgba(255, 255, 255, 0.8)'
-    this.previewContainer.style.boxShadow = '0 0 8px rgba(255, 255, 255, 0.5)'
-
-    setTimeout(() => {
-      this.previewContainer.style.border = originalBorder
-      this.previewContainer.style.boxShadow = originalBoxShadow
-    }, 100)
-  }
-
-  updatePreviewRender() {
-    if (!this.previewRenderer || !this.previewContainer || !this.showPreview)
-      return
-
-    if (
-      !this.previewCamera ||
-      (this.activeCamera instanceof THREE.PerspectiveCamera &&
-        !(this.previewCamera instanceof THREE.PerspectiveCamera)) ||
-      (this.activeCamera instanceof THREE.OrthographicCamera &&
-        !(this.previewCamera instanceof THREE.OrthographicCamera))
-    ) {
-      this.previewCamera = this.activeCamera.clone()
-    }
-
-    this.previewCamera.position.copy(this.activeCamera.position)
-    this.previewCamera.rotation.copy(this.activeCamera.rotation)
-
-    const aspect = this.targetWidth / this.targetHeight
-
-    if (this.activeCamera instanceof THREE.OrthographicCamera) {
-      const activeOrtho = this.activeCamera as THREE.OrthographicCamera
-      const previewOrtho = this.previewCamera as THREE.OrthographicCamera
-
-      const frustumHeight =
-        (activeOrtho.top - activeOrtho.bottom) / activeOrtho.zoom
-
-      const frustumWidth = frustumHeight * aspect
-
-      previewOrtho.top = frustumHeight / 2
-      previewOrtho.left = -frustumWidth / 2
-      previewOrtho.right = frustumWidth / 2
-      previewOrtho.bottom = -frustumHeight / 2
-      previewOrtho.zoom = 1
-
-      previewOrtho.updateProjectionMatrix()
-    } else {
-      ;(this.previewCamera as THREE.PerspectiveCamera).aspect = aspect
-      ;(this.previewCamera as THREE.PerspectiveCamera).fov = (
-        this.activeCamera as THREE.PerspectiveCamera
-      ).fov
-      ;(this.previewCamera as THREE.PerspectiveCamera).updateProjectionMatrix()
-    }
-
-    this.previewCamera.lookAt(this.controls.target)
-
-    const previewHeight =
-      (this.previewWidth * this.targetHeight) / this.targetWidth
-    this.previewRenderer.setSize(this.previewWidth, previewHeight, false)
-
-    this.previewRenderer.outputColorSpace = THREE.SRGBColorSpace
-
-    this.previewRenderer.clear()
-
-    if (this.previewBackgroundMesh && this.previewBackgroundTexture) {
-      const material = this.previewBackgroundMesh
-        .material as THREE.MeshBasicMaterial
-      if (material.map) {
-        const currentToneMapping = this.previewRenderer.toneMapping
-        const currentExposure = this.previewRenderer.toneMappingExposure
-
-        this.previewRenderer.toneMapping = THREE.NoToneMapping
-
-        this.previewRenderer.render(
-          this.previewBackgroundScene,
-          this.previewBackgroundCamera
-        )
-
-        this.previewRenderer.toneMapping = currentToneMapping
-        this.previewRenderer.toneMappingExposure = currentExposure
-      }
-    }
-
-    this.previewRenderer.render(this.scene, this.previewCamera)
-  }
-
-  updatePreviewSize() {
-    if (!this.previewContainer) return
-
-    const previewHeight =
-      (this.previewWidth * this.targetHeight) / this.targetWidth
-
-    this.previewRenderer?.setSize(this.previewWidth, previewHeight, false)
-  }
-
-  setTargetSize(width: number, height: number) {
-    const oldAspect = this.targetWidth / this.targetHeight
-
-    this.targetWidth = width
-    this.targetHeight = height
-
-    this.updatePreviewSize()
-
-    const newAspect = width / height
-    if (Math.abs(oldAspect - newAspect) > 0.001) {
-      this.updateBackgroundSize(
-        this.previewBackgroundTexture,
-        this.previewBackgroundMesh,
-        width,
-        height
-      )
-    }
-
-    if (this.previewRenderer && this.previewCamera) {
-      if (this.previewCamera instanceof THREE.PerspectiveCamera) {
-        this.previewCamera.aspect = width / height
-        this.previewCamera.updateProjectionMatrix()
-      } else if (this.previewCamera instanceof THREE.OrthographicCamera) {
-        const frustumSize = 10
-        const aspect = width / height
-        this.previewCamera.left = (-frustumSize * aspect) / 2
-        this.previewCamera.right = (frustumSize * aspect) / 2
-        this.previewCamera.updateProjectionMatrix()
-      }
-    }
-  }
-
-  createViewHelper(container: Element | HTMLElement) {
-    this.viewHelperContainer = document.createElement('div')
-
-    this.viewHelperContainer.style.position = 'absolute'
-    this.viewHelperContainer.style.bottom = '0'
-    this.viewHelperContainer.style.left = '0'
-    this.viewHelperContainer.style.width = '128px'
-    this.viewHelperContainer.style.height = '128px'
-    this.viewHelperContainer.addEventListener('pointerup', (event) => {
-      event.stopPropagation()
-
-      this.viewHelper.handleClick(event)
-    })
-
-    this.viewHelperContainer.addEventListener('pointerdown', (event) => {
-      event.stopPropagation()
-    })
-
-    container.appendChild(this.viewHelperContainer)
-
-    this.viewHelper = new ViewHelper(
-      this.activeCamera,
-      this.viewHelperContainer
-    )
-
-    this.viewHelper.center = this.controls.target
-  }
-
-  setFOV(fov: number) {
-    if (this.activeCamera === this.perspectiveCamera) {
-      this.perspectiveCamera.fov = fov
-      this.perspectiveCamera.updateProjectionMatrix()
-      this.renderer.render(this.scene, this.activeCamera)
-    }
-
-    if (
-      this.previewRenderer &&
-      this.previewCamera instanceof THREE.PerspectiveCamera
-    ) {
-      this.previewCamera.fov = fov
-      this.previewCamera.updateProjectionMatrix()
-      this.previewRenderer.render(this.scene, this.previewCamera)
-    }
-
-    this.emitEvent('fovChange', fov)
-  }
-
-  getCameraState() {
-    const currentType = this.getCurrentCameraType()
-    return {
-      position: this.activeCamera.position.clone(),
-      target: this.controls.target.clone(),
-      zoom:
-        this.activeCamera instanceof THREE.OrthographicCamera
-          ? this.activeCamera.zoom
-          : (this.activeCamera as THREE.PerspectiveCamera).zoom,
-      cameraType: currentType
-    }
-  }
-
-  setCameraState(state: {
-    position: THREE.Vector3
-    target: THREE.Vector3
-    zoom: number
-    cameraType: 'perspective' | 'orthographic'
-  }) {
-    this.activeCamera.position.copy(state.position)
-
-    this.controls.target.copy(state.target)
-
-    if (this.activeCamera instanceof THREE.OrthographicCamera) {
-      this.activeCamera.zoom = state.zoom
-      this.activeCamera.updateProjectionMatrix()
-    } else if (this.activeCamera instanceof THREE.PerspectiveCamera) {
-      this.activeCamera.zoom = state.zoom
-      this.activeCamera.updateProjectionMatrix()
-    }
-
-    this.controls.update()
-  }
-
-  setUpDirection(
-    direction: 'original' | '-x' | '+x' | '-y' | '+y' | '-z' | '+z'
-  ) {
-    if (!this.currentModel) return
-
-    if (!this.originalRotation && this.currentModel.rotation) {
-      this.originalRotation = this.currentModel.rotation.clone()
-    }
-
-    this.currentUpDirection = direction
-
-    if (this.originalRotation) {
-      this.currentModel.rotation.copy(this.originalRotation)
-    }
-
-    switch (direction) {
-      case 'original':
-        break
-      case '-x':
-        this.currentModel.rotation.z = Math.PI / 2
-        break
-      case '+x':
-        this.currentModel.rotation.z = -Math.PI / 2
-        break
-      case '-y':
-        this.currentModel.rotation.x = Math.PI
-        break
-      case '+y':
-        break
-      case '-z':
-        this.currentModel.rotation.x = Math.PI / 2
-        break
-      case '+z':
-        this.currentModel.rotation.x = -Math.PI / 2
-        break
-    }
-
-    this.renderer.render(this.scene, this.activeCamera)
-  }
-
-  setMaterialMode(mode: 'original' | 'normal' | 'wireframe' | 'depth') {
-    this.materialMode = mode
-
-    if (this.currentModel) {
-      if (mode === 'depth') {
-        this.renderer.outputColorSpace = THREE.LinearSRGBColorSpace
-      } else {
-        this.renderer.outputColorSpace = THREE.SRGBColorSpace
-      }
-
-      this.currentModel.traverse((child) => {
-        if (child instanceof THREE.Mesh) {
-          switch (mode) {
-            case 'depth':
-              if (!this.originalMaterials.has(child)) {
-                this.originalMaterials.set(child, child.material)
-              }
-              const depthMat = new THREE.MeshDepthMaterial({
-                depthPacking: THREE.BasicDepthPacking,
-                side: THREE.DoubleSide
-              })
-
-              depthMat.onBeforeCompile = (shader) => {
-                shader.uniforms.cameraType = {
-                  value:
-                    this.activeCamera instanceof THREE.OrthographicCamera
-                      ? 1.0
-                      : 0.0
-                }
-
-                shader.fragmentShader = `
-                  uniform float cameraType;
-                  ${shader.fragmentShader}
-                `
-
-                shader.fragmentShader = shader.fragmentShader.replace(
-                  /gl_FragColor\s*=\s*vec4\(\s*vec3\(\s*1.0\s*-\s*fragCoordZ\s*\)\s*,\s*opacity\s*\)\s*;/,
-                  `
-                    float depth = 1.0 - fragCoordZ;
-                    if (cameraType > 0.5) {
-                      depth = pow(depth, 400.0);
-                    } else {
-                      depth = pow(depth, 0.6);
-                    }
-                    gl_FragColor = vec4(vec3(depth), opacity);
-                  `
-                )
-              }
-
-              depthMat.customProgramCacheKey = () => {
-                return this.activeCamera instanceof THREE.OrthographicCamera
-                  ? 'ortho'
-                  : 'persp'
-              }
-
-              child.material = depthMat
-              break
-            case 'normal':
-              if (!this.originalMaterials.has(child)) {
-                this.originalMaterials.set(child, child.material)
-              }
-              child.material = new THREE.MeshNormalMaterial({
-                flatShading: false,
-                side: THREE.DoubleSide,
-                normalScale: new THREE.Vector2(1, 1),
-                transparent: false,
-                opacity: 1.0
-              })
-              child.geometry.computeVertexNormals()
-              break
-
-            case 'wireframe':
-              if (!this.originalMaterials.has(child)) {
-                this.originalMaterials.set(child, child.material)
-              }
-              child.material = new THREE.MeshBasicMaterial({
-                color: 0xffffff,
-                wireframe: true,
-                transparent: false,
-                opacity: 1.0
-              })
-              break
-
-            case 'original':
-              const originalMaterial = this.originalMaterials.get(child)
-              if (originalMaterial) {
-                child.material = originalMaterial
-              } else {
-                child.material = this.standardMaterial
-              }
-              break
-          }
-        }
-      })
-
-      this.renderer.render(this.scene, this.activeCamera)
-
-      this.emitEvent('materialModeChange', mode)
-    }
-  }
-
-  setupLights() {
-    const ambientLight = new THREE.AmbientLight(0xffffff, 0.5)
-    this.scene.add(ambientLight)
-    this.lights.push(ambientLight)
-
-    const mainLight = new THREE.DirectionalLight(0xffffff, 0.8)
-    mainLight.position.set(0, 10, 10)
-    this.scene.add(mainLight)
-    this.lights.push(mainLight)
-
-    const backLight = new THREE.DirectionalLight(0xffffff, 0.5)
-    backLight.position.set(0, 10, -10)
-    this.scene.add(backLight)
-    this.lights.push(backLight)
-
-    const leftFillLight = new THREE.DirectionalLight(0xffffff, 0.3)
-    leftFillLight.position.set(-10, 0, 0)
-    this.scene.add(leftFillLight)
-    this.lights.push(leftFillLight)
-
-    const rightFillLight = new THREE.DirectionalLight(0xffffff, 0.3)
-    rightFillLight.position.set(10, 0, 0)
-    this.scene.add(rightFillLight)
-    this.lights.push(rightFillLight)
-
-    const bottomLight = new THREE.DirectionalLight(0xffffff, 0.2)
-    bottomLight.position.set(0, -10, 0)
-    this.scene.add(bottomLight)
-    this.lights.push(bottomLight)
-  }
-
-  toggleCamera(cameraType?: 'perspective' | 'orthographic') {
-    const oldCamera = this.activeCamera
-
-    const position = oldCamera.position.clone()
-    const rotation = oldCamera.rotation.clone()
-    const target = this.controls.target.clone()
-
-    if (!cameraType) {
-      this.activeCamera =
-        oldCamera === this.perspectiveCamera
-          ? this.orthographicCamera
-          : this.perspectiveCamera
-    } else {
-      this.activeCamera =
-        cameraType === 'perspective'
-          ? this.perspectiveCamera
-          : this.orthographicCamera
-
-      if (oldCamera === this.activeCamera) {
-        return
-      }
-    }
-
-    if (this.previewCamera) {
-      this.previewCamera = null
-    }
-    this.previewCamera = this.activeCamera.clone()
-
-    this.activeCamera.position.copy(position)
-    this.activeCamera.rotation.copy(rotation)
-
-    if (this.materialMode === 'depth' && oldCamera !== this.activeCamera) {
-      this.setMaterialMode('depth')
-    }
-
-    this.controls.object = this.activeCamera
-    this.controls.target.copy(target)
-    this.controls.update()
-
-    this.viewHelper.dispose()
-    this.viewHelper = new ViewHelper(
-      this.activeCamera,
-      this.viewHelperContainer
-    )
-    this.viewHelper.center = this.controls.target
-
-    this.handleResize()
-    this.updatePreviewRender()
-
-    this.emitEvent('cameraTypeChange', cameraType)
-  }
-
-  getCurrentCameraType(): 'perspective' | 'orthographic' {
-    return this.activeCamera === this.perspectiveCamera
-      ? 'perspective'
-      : 'orthographic'
-  }
-
-  toggleGrid(showGrid: boolean) {
-    if (this.gridHelper) {
-      this.gridHelper.visible = showGrid
-    }
-
-    this.emitEvent('showGridChange', showGrid)
-  }
-
-  togglePreview(showPreview: boolean) {
-    if (this.previewRenderer) {
-      this.showPreview = showPreview
-
-      this.previewContainer.style.display = this.showPreview ? 'block' : 'none'
-    }
-
-    this.emitEvent('showPreviewChange', showPreview)
-  }
-
-  setLightIntensity(intensity: number) {
-    this.lights.forEach((light) => {
-      if (light instanceof THREE.DirectionalLight) {
-        if (light === this.lights[1]) {
-          light.intensity = intensity * 0.8
-        } else if (light === this.lights[2]) {
-          light.intensity = intensity * 0.5
-        } else if (light === this.lights[5]) {
-          light.intensity = intensity * 0.2
-        } else {
-          light.intensity = intensity * 0.3
-        }
-      } else if (light instanceof THREE.AmbientLight) {
-        light.intensity = intensity * 0.5
-      }
-    })
-
-    this.emitEvent('lightIntensityChange', intensity)
-  }
-
-  startAnimation() {
+  private startAnimation(): void {
     const animate = () => {
       this.animationFrameId = requestAnimationFrame(animate)
 
-      if (this.showPreview) {
-        this.updatePreviewRender()
+      if (this.previewManager.showPreview) {
+        this.previewManager.updatePreviewRender()
       }
 
       const delta = this.clock.getDelta()
-
-      if (this.viewHelper.animating) {
-        this.viewHelper.update(delta)
-
-        if (!this.viewHelper.animating) {
-          this.storeNodeProperty('Camera Info', this.getCameraState())
-        }
-      }
+      this.viewHelperManager.update(delta)
+      this.controlsManager.update()
 
       this.renderer.clear()
+      this.sceneManager.renderBackground()
+      this.renderer.render(
+        this.sceneManager.scene,
+        this.cameraManager.activeCamera
+      )
 
-      if (this.backgroundMesh && this.backgroundTexture) {
-        const material = this.backgroundMesh.material as THREE.MeshBasicMaterial
-        if (material.map) {
-          const currentToneMapping = this.renderer.toneMapping
-          const currentExposure = this.renderer.toneMappingExposure
-
-          this.renderer.toneMapping = THREE.NoToneMapping
-
-          this.renderer.render(this.backgroundScene, this.backgroundCamera)
-
-          this.renderer.toneMapping = currentToneMapping
-          this.renderer.toneMappingExposure = currentExposure
-        }
+      if (this.viewHelperManager.viewHelper.render) {
+        this.viewHelperManager.viewHelper.render(this.renderer)
       }
-
-      this.controls.update()
-      this.renderer.render(this.scene, this.activeCamera)
-      this.viewHelper.render(this.renderer)
     }
+
     animate()
   }
 
-  clearModel() {
-    const objectsToRemove: THREE.Object3D[] = []
-
-    this.scene.traverse((object) => {
-      const isEnvironmentObject =
-        object === this.gridHelper ||
-        this.lights.includes(object as THREE.Light) ||
-        object === this.perspectiveCamera ||
-        object === this.orthographicCamera
-
-      if (!isEnvironmentObject) {
-        objectsToRemove.push(object)
-      }
-    })
-
-    objectsToRemove.forEach((obj) => {
-      if (obj.parent && obj.parent !== this.scene) {
-        obj.parent.remove(obj)
-      } else {
-        this.scene.remove(obj)
-      }
-
-      if (obj instanceof THREE.Mesh) {
-        obj.geometry?.dispose()
-        if (Array.isArray(obj.material)) {
-          obj.material.forEach((material) => material.dispose())
-        } else {
-          obj.material?.dispose()
-        }
-      }
-    })
-
-    this.resetScene()
-  }
-
-  protected resetScene() {
-    this.currentModel = null
-    this.originalRotation = null
-
-    const defaultDistance = 10
-    this.perspectiveCamera.position.set(
-      defaultDistance,
-      defaultDistance,
-      defaultDistance
+  setBackgroundColor(color: string): void {
+    this.sceneManager.setBackgroundColor(color)
+    this.renderer.render(
+      this.sceneManager.scene,
+      this.cameraManager.activeCamera
     )
-    this.orthographicCamera.position.set(
-      defaultDistance,
-      defaultDistance,
-      defaultDistance
-    )
-
-    this.perspectiveCamera.lookAt(0, 0, 0)
-    this.orthographicCamera.lookAt(0, 0, 0)
-
-    const frustumSize = 10
-    const aspect =
-      this.renderer.domElement.width / this.renderer.domElement.height
-    this.orthographicCamera.left = (-frustumSize * aspect) / 2
-    this.orthographicCamera.right = (frustumSize * aspect) / 2
-    this.orthographicCamera.top = frustumSize / 2
-    this.orthographicCamera.bottom = -frustumSize / 2
-
-    this.perspectiveCamera.updateProjectionMatrix()
-    this.orthographicCamera.updateProjectionMatrix()
-
-    this.controls.target.set(0, 0, 0)
-    this.controls.update()
-
-    this.renderer.render(this.scene, this.activeCamera)
-
-    this.materialMode = 'original'
-    this.originalMaterials = new WeakMap()
-    this.renderer.outputColorSpace = THREE.SRGBColorSpace
   }
 
-  remove() {
-    if (this.animationFrameId !== null) {
-      cancelAnimationFrame(this.animationFrameId)
-    }
+  async setBackgroundImage(uploadPath: string): Promise<void> {
+    await this.sceneManager.setBackgroundImage(uploadPath)
 
-    if (this.backgroundTexture) {
-      this.backgroundTexture.dispose()
-    }
-
-    if (this.previewBackgroundTexture) {
-      this.previewBackgroundTexture.dispose()
-    }
-
-    if (this.backgroundMesh) {
-      this.backgroundMesh.geometry.dispose()
-      ;(this.backgroundMesh.material as THREE.Material).dispose()
-    }
-
-    if (this.previewBackgroundMesh) {
-      this.previewBackgroundMesh.geometry.dispose()
-      ;(this.previewBackgroundMesh.material as THREE.Material).dispose()
-    }
-
-    this.controls.dispose()
-    this.viewHelper.dispose()
-    this.renderer.dispose()
-    this.renderer.domElement.remove()
-    this.scene.clear()
-  }
-
-  protected async loadModelInternal(
-    url: string,
-    fileExtension: string
-  ): Promise<THREE.Object3D | null> {
-    let model: THREE.Object3D | null = null
-
-    switch (fileExtension) {
-      case 'stl':
-        const geometry = await this.stlLoader.loadAsync(url)
-
-        this.originalModel = geometry
-
-        geometry.computeVertexNormals()
-        const mesh = new THREE.Mesh(geometry, this.standardMaterial)
-        const group = new THREE.Group()
-        group.add(mesh)
-        model = group
-        break
-
-      case 'fbx':
-        const fbxModel = await this.fbxLoader.loadAsync(url)
-
-        this.originalModel = fbxModel
-
-        model = fbxModel
-
-        fbxModel.traverse((child) => {
-          if (child instanceof THREE.Mesh) {
-            this.originalMaterials.set(child, child.material)
-          }
-        })
-
-        break
-
-      case 'obj':
-        if (this.materialMode === 'original') {
-          const mtlUrl = url.replace(/\.obj([^.]*$)/, '.mtl$1')
-          try {
-            const materials = await this.mtlLoader.loadAsync(mtlUrl)
-            materials.preload()
-            this.objLoader.setMaterials(materials)
-          } catch (e) {
-            console.log(
-              'No MTL file found or error loading it, continuing without materials'
-            )
-          }
-        }
-        model = await this.objLoader.loadAsync(url)
-        model.traverse((child) => {
-          if (child instanceof THREE.Mesh) {
-            this.originalMaterials.set(child, child.material)
-          }
-        })
-        break
-
-      case 'gltf':
-      case 'glb':
-        const gltf = await this.gltfLoader.loadAsync(url)
-
-        this.originalModel = gltf
-
-        model = gltf.scene
-        gltf.scene.traverse((child) => {
-          if (child instanceof THREE.Mesh) {
-            child.geometry.computeVertexNormals()
-            this.originalMaterials.set(child, child.material)
-          }
-        })
-        break
-    }
-
-    return model
-  }
-
-  async loadModel(url: string, originalFileName?: string) {
-    try {
-      this.clearModel()
-
-      let fileExtension: string | undefined
-      if (originalFileName) {
-        fileExtension = originalFileName.split('.').pop()?.toLowerCase()
-      } else {
-        const filename = new URLSearchParams(url.split('?')[1]).get('filename')
-        fileExtension = filename?.split('.').pop()?.toLowerCase()
-      }
-
-      if (!fileExtension) {
-        useToastStore().addAlert('Could not determine file type')
-        return
-      }
-
-      let model = await this.loadModelInternal(url, fileExtension)
-
-      if (model) {
-        this.currentModel = model
-        await this.setupModel(model)
-      }
-    } catch (error) {
-      console.error('Error loading model:', error)
+    if (this.previewManager.previewRenderer) {
+      this.previewManager.updateBackgroundTexture(
+        this.sceneManager.backgroundTexture
+      )
     }
   }
 
-  protected async setupModel(model: THREE.Object3D) {
-    const box = new THREE.Box3().setFromObject(model)
-    const size = box.getSize(new THREE.Vector3())
-    const center = box.getCenter(new THREE.Vector3())
+  removeBackgroundImage(): void {
+    this.sceneManager.removeBackgroundImage()
 
-    const maxDim = Math.max(size.x, size.y, size.z)
-    const targetSize = 5
-    const scale = targetSize / maxDim
-    model.scale.multiplyScalar(scale)
-
-    box.setFromObject(model)
-    box.getCenter(center)
-    box.getSize(size)
-
-    model.position.set(-center.x, -box.min.y, -center.z)
-
-    this.scene.add(model)
-
-    if (this.materialMode !== 'original') {
-      this.setMaterialMode(this.materialMode)
+    if (
+      this.previewManager.previewRenderer &&
+      this.previewManager.previewCamera
+    ) {
+      this.previewManager.updateBackgroundTexture(null)
     }
-
-    if (this.currentUpDirection !== 'original') {
-      this.setUpDirection(this.currentUpDirection)
-    }
-
-    await this.setupCamera(size)
   }
 
-  protected async setupCamera(size: THREE.Vector3) {
-    const distance = Math.max(size.x, size.z) * 2
-    const height = size.y * 2
+  toggleGrid(showGrid: boolean): void {
+    this.sceneManager.toggleGrid(showGrid)
+  }
 
-    this.perspectiveCamera.position.set(distance, height, distance)
-    this.orthographicCamera.position.set(distance, height, distance)
+  toggleCamera(cameraType?: 'perspective' | 'orthographic'): void {
+    this.cameraManager.toggleCamera(cameraType)
 
-    if (this.activeCamera === this.perspectiveCamera) {
-      this.perspectiveCamera.lookAt(0, size.y / 2, 0)
-      this.perspectiveCamera.updateProjectionMatrix()
-    } else {
-      const frustumSize = Math.max(size.x, size.y, size.z) * 2
-      const aspect =
-        this.renderer.domElement.width / this.renderer.domElement.height
-      this.orthographicCamera.left = (-frustumSize * aspect) / 2
-      this.orthographicCamera.right = (frustumSize * aspect) / 2
-      this.orthographicCamera.top = frustumSize / 2
-      this.orthographicCamera.bottom = -frustumSize / 2
-      this.orthographicCamera.lookAt(0, size.y / 2, 0)
-      this.orthographicCamera.updateProjectionMatrix()
-    }
-
-    this.controls.target.set(0, size.y / 2, 0)
-    this.controls.update()
-
-    this.renderer.outputColorSpace = THREE.SRGBColorSpace
-    this.renderer.toneMapping = THREE.ACESFilmicToneMapping
-    this.renderer.toneMappingExposure = 1
+    this.controlsManager.updateCamera(this.cameraManager.activeCamera)
+    this.viewHelperManager.recreateViewHelper()
 
     this.handleResize()
   }
 
-  refreshViewport() {
+  getCurrentCameraType(): 'perspective' | 'orthographic' {
+    return this.cameraManager.getCurrentCameraType()
+  }
+
+  setCameraState(state: CameraState): void {
+    this.cameraManager.setCameraState(state)
+  }
+
+  getCameraState(): CameraState {
+    return this.cameraManager.getCameraState()
+  }
+
+  setFOV(fov: number): void {
+    this.cameraManager.setFOV(fov)
+    this.renderer.render(
+      this.sceneManager.scene,
+      this.cameraManager.activeCamera
+    )
+  }
+
+  setMaterialMode(mode: MaterialMode): void {
+    this.modelManager.setMaterialMode(mode)
+
+    this.renderer.render(
+      this.sceneManager.scene,
+      this.cameraManager.activeCamera
+    )
+  }
+
+  async loadModel(url: string, originalFileName?: string): Promise<void> {
+    this.cameraManager.reset()
+    this.controlsManager.reset()
+    this.modelManager.reset()
+
+    await this.loaderManager.loadModel(url, originalFileName)
+
     this.handleResize()
   }
 
-  handleResize() {
+  clearModel(): void {
+    this.modelManager.clearModel()
+  }
+
+  setUpDirection(direction: UpDirection): void {
+    this.modelManager.setUpDirection(direction)
+    this.renderer.render(
+      this.sceneManager.scene,
+      this.cameraManager.activeCamera
+    )
+  }
+
+  setLightIntensity(intensity: number): void {
+    this.lightingManager.setLightIntensity(intensity)
+  }
+
+  togglePreview(showPreview: boolean): void {
+    this.previewManager.togglePreview(showPreview)
+  }
+
+  setTargetSize(width: number, height: number): void {
+    this.previewManager.setTargetSize(width, height)
+  }
+
+  addEventListener(event: string, callback: (data?: any) => void): void {
+    this.eventManager.addEventListener(event, callback)
+  }
+
+  removeEventListener(event: string, callback: (data?: any) => void): void {
+    this.eventManager.removeEventListener(event, callback)
+  }
+
+  refreshViewport(): void {
+    this.handleResize()
+  }
+
+  handleResize(): void {
     const parentElement = this.renderer?.domElement?.parentElement
 
     if (!parentElement) {
@@ -1190,131 +292,41 @@ class Load3d {
       return
     }
 
-    const width = parentElement?.clientWidth
-    const height = parentElement?.clientHeight
+    const width = parentElement.clientWidth
+    const height = parentElement.clientHeight
 
-    if (this.activeCamera === this.perspectiveCamera) {
-      this.perspectiveCamera.aspect = width / height
-      this.perspectiveCamera.updateProjectionMatrix()
-    } else {
-      const frustumSize = 10
-      const aspect = width / height
-      this.orthographicCamera.left = (-frustumSize * aspect) / 2
-      this.orthographicCamera.right = (frustumSize * aspect) / 2
-      this.orthographicCamera.top = frustumSize / 2
-      this.orthographicCamera.bottom = -frustumSize / 2
-      this.orthographicCamera.updateProjectionMatrix()
-    }
+    this.cameraManager.handleResize(width, height)
+    this.sceneManager.handleResize(width, height)
 
     this.renderer.setSize(width, height)
 
-    if (this.backgroundTexture && this.backgroundMesh) {
-      this.updateBackgroundSize(
-        this.backgroundTexture,
-        this.backgroundMesh,
-        width,
-        height
-      )
+    this.previewManager.handleResize()
+  }
+
+  captureScene(width: number, height: number): Promise<CaptureResult> {
+    return this.sceneManager.captureScene(width, height)
+  }
+
+  loadNodeProperty(name: string, defaultValue: any) {
+    return this.nodeStorage.loadNodeProperty(name, defaultValue)
+  }
+
+  remove(): void {
+    if (this.animationFrameId !== null) {
+      cancelAnimationFrame(this.animationFrameId)
     }
 
-    this.setTargetSize(this.targetWidth, this.targetHeight)
-  }
+    this.sceneManager.dispose()
+    this.cameraManager.dispose()
+    this.controlsManager.dispose()
+    this.lightingManager.dispose()
+    this.viewHelperManager.dispose()
+    this.previewManager.dispose()
+    this.loaderManager.dispose()
+    this.modelManager.dispose()
 
-  animate = () => {
-    requestAnimationFrame(this.animate)
-
-    this.controls.update()
-    this.renderer.render(this.scene, this.activeCamera)
-  }
-
-  captureScene(
-    width: number,
-    height: number
-  ): Promise<{ scene: string; mask: string }> {
-    return new Promise(async (resolve, reject) => {
-      try {
-        this.updatePreviewSize()
-        const originalWidth = this.renderer.domElement.width
-        const originalHeight = this.renderer.domElement.height
-        const originalClearColor = this.renderer.getClearColor(
-          new THREE.Color()
-        )
-        const originalClearAlpha = this.renderer.getClearAlpha()
-        const originalToneMapping = this.renderer.toneMapping
-        const originalExposure = this.renderer.toneMappingExposure
-
-        this.renderer.setSize(width, height)
-
-        if (this.activeCamera === this.perspectiveCamera) {
-          this.perspectiveCamera.aspect = width / height
-          this.perspectiveCamera.updateProjectionMatrix()
-        } else {
-          const frustumSize = 10
-          const aspect = width / height
-          this.orthographicCamera.left = (-frustumSize * aspect) / 2
-          this.orthographicCamera.right = (frustumSize * aspect) / 2
-          this.orthographicCamera.top = frustumSize / 2
-          this.orthographicCamera.bottom = -frustumSize / 2
-          this.orthographicCamera.updateProjectionMatrix()
-        }
-
-        if (this.backgroundTexture && this.backgroundMesh) {
-          this.updateBackgroundSize(
-            this.backgroundTexture,
-            this.backgroundMesh,
-            width,
-            height
-          )
-        }
-
-        this.renderer.clear()
-
-        if (this.backgroundMesh && this.backgroundTexture) {
-          const material = this.backgroundMesh
-            .material as THREE.MeshBasicMaterial
-
-          if (material.map) {
-            this.renderer.toneMapping = THREE.NoToneMapping
-            this.renderer.render(this.backgroundScene, this.backgroundCamera)
-            this.renderer.toneMapping = originalToneMapping
-            this.renderer.toneMappingExposure = originalExposure
-          }
-        }
-
-        this.renderer.render(this.scene, this.activeCamera)
-        const sceneData = this.renderer.domElement.toDataURL('image/png')
-
-        this.renderer.setClearColor(0x000000, 0)
-        this.renderer.clear()
-        this.renderer.render(this.scene, this.activeCamera)
-        const maskData = this.renderer.domElement.toDataURL('image/png')
-
-        this.renderer.setClearColor(originalClearColor, originalClearAlpha)
-        this.renderer.setSize(originalWidth, originalHeight)
-        this.handleResize()
-
-        resolve({ scene: sceneData, mask: maskData })
-      } catch (error) {
-        reject(error)
-      }
-    })
-  }
-
-  createSTLMaterial() {
-    return new THREE.MeshStandardMaterial({
-      color: 0x808080,
-      metalness: 0.1,
-      roughness: 0.8,
-      flatShading: false,
-      side: THREE.DoubleSide
-    })
-  }
-
-  setBackgroundColor(color: string) {
-    this.renderer.setClearColor(new THREE.Color(color))
-    this.renderer.render(this.scene, this.activeCamera)
-
-    this.emitEvent('backgroundColorChange', color)
+    this.renderer.dispose()
+    this.renderer.domElement.remove()
   }
 }
