+++ src/locales/en/main.json
@@ -847,6 +847,7 @@
     "fov": "FOV",
     "previewOutput": "Preview Output",
     "uploadBackgroundImage": "Upload Background Image",
-    "removeBackgroundImage": "Remove Background Image"
+    "removeBackgroundImage": "Remove Background Image",
+    "loadingModel": "Loading 3D Model..."
   }
 }
